updating app.js + home.js
