+++ client/src/Home.js
@@ -1,19 +1,24 @@
 
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import Search from './components/Search';
 import Results from './components/Results';
 import Popup from './components/Popup';
 import GNI from './components/GNI.png';
+import app from "./base";
 
-const Home = () => {
+
+
+console.log(process.env.REACT_APP_API_KEY);
+
+function Home() {
   const [state, setState] = useState({
     searchInput: "",
     results: [],
     selected: {}
   });
 
-  const apiurl = 'http://www.omdbapi.com/?apikey=6d71121d';
+  const apiurl = `http://www.omdbapi.com/?apikey=${process.env.REACT_APP_API_KEY}`;
 
   const search = (e) => {
     if (e.key === "Enter") {
@@ -25,6 +30,22 @@ const Home = () => {
         })
       });
     }
+  }
+
+  const getMovies = () => {
+    axios("/api/movies").then(({ data }) => {
+      let results = data;
+      console.log(data);
+      let movies = data.forEach(({ Title }) => {
+        axios(apiurl + "&s=" + Title).then(({ data }) => {
+          let results = data.Search;
+
+          setState(prevState => {
+            return { ...prevState, results: [...prevState.results, results[0]] }
+          })
+        });
+      })
+    });
   }
 
   const handleInput = (e) => {
@@ -52,26 +73,30 @@ const Home = () => {
       return { ...prevState, selected: {} }
     });
   }
+
+  useEffect(() => { getMovies() }, [])
+
   return (
-    <>
-      <div className="App">
-        <header>
-          <div className='hero'>
-            <img id='GNI' src={GNI} alt="Girl's Night In Neon" />
-          </div>
-          <h1>Search For A Movie!</h1>
-        </header>
-        <main>
-          <Search handleInput={handleInput} search={search} />
 
-          <Results results={state.results} openPopup={openPopup} />
+    <div className="App">
+      <h1>Home</h1>
+      <button onClick={() => app.auth().signOut()}>Sign out</button>
+      <header>
+        <div className='hero'>
+          <img id='GNI' src={GNI} alt="Girl's Night In Neon" />
+        </div>
+        <h1>Search For A Movie!</h1>
+      </header>
+      <main>
+        <Search handleInput={handleInput} search={search} />
 
-          {(typeof state.selected.Title !== "undefined") ? <Popup selected={state.selected} closePopup={closePopup} /> : false}
+        <Results results={state.results} openPopup={openPopup} />
 
-        </main>
-      </div>
-    </>
+        {(typeof state.selected.Title !== "undefined") ? <Popup selected={state.selected} closePopup={closePopup} /> : false}
+
+      </main>
+    </div>
   );
-};
+}
 
 export default Home;
